requirements.txt
fastapi
uvicorn
scikit-learn
pandas
joblib
python-multipart
python-dotenv 